+++ unicorn_binance_websocket_api/unicorn_binance_websocket_api_socket.py
@@ -146,6 +146,11 @@ class BinanceWebSocketApiSocket(object):
                             stream_buffer_name = self.manager.stream_list[self.stream_id]['stream_buffer_name']
                         except KeyError:
                             stream_buffer_name = False
+                        if self.manager.stream_list[self.stream_id]['last_received_data_record'] is None:
+                            self.manager.add_to_stream_signal_buffer("FIRST_RECEIVED_DATA",
+                                                                     stream_id,
+                                                                     received_stream_data)
+                        self.manager.stream_list[self.stream_id]['last_received_data_record'] = received_stream_data
                         self.manager.process_stream_data(received_stream_data,
                                                          stream_buffer_name=stream_buffer_name)
                         if "error" in received_stream_data_json: